requirements.txt
flask
open_clip_torch
numpy 
Pillow 
tqdm
pandas
matplotlib 
scikit-learn
